infra/pyproject.toml
[project]
name = "infra"
version = "0.1.0"
description = "AWS CDK infrastructure for python-aws-monorepo-boilerplate"
requires-python = ">=3.13"
dependencies = [
    "aws-cdk-lib>=2.170.0",
    "constructs>=10.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["stacks", "utils"]
